requirements.txt
streamlit>=1.30.0
ollama>=0.1.0
python-dotenv>=1.0.0


